src/indexer/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-functions
azure-search-documents
httpx
langchain
openai
PyMuPDF
python-dotenv
